Refactor CLI into modular package
