+++ src/egregora/cli/commands/__init__.py
@@ -1,0 +1,13 @@
+"""Command modules for the Egregora CLI."""
+
+from __future__ import annotations
+
+from . import doctor, init, runs, views, write
+
+__all__ = [
+    "doctor",
+    "init",
+    "runs",
+    "views",
+    "write",
+]
